django/pyproject.toml
[project]
name = "django"
version = "0.0.0-dev"
dependencies = [
    "Django<4.3",
    "bcrypt",
    "django-behaviors",
    "django-environ",
    "django-ipware",
    "django-axes",
    "whitenoise",
    "django-storages",
    "djangorestframework",
    "djangorestframework-camel-case",
    "drf-jwt",
    "drf-spectacular[sidecar]",
    "django-filter",
    "django-split-settings",
    "django-healthchecks",
    "redis",
    "sentry-sdk",
    "Pillow",
    "psycopg2-binary",
]


[project.optional-dependencies]
dev = [
    "ipython",

    "pytest-django>=3.9",
    "pytest-deadfixtures",
    "pytest-env",
    "pytest-freezegun",
    "pytest-mock",
    "pytest-randomly",

    "black",

    "autoflake==1.7.0",

    "dotenv-linter",

    "freezegun",
    "mixer",

    "jedi",
    "flake8-absolute-import",
    "flake8-black",
    "flake8-bugbear",
    "flake8-cognitive-complexity",
    "flake8-django",
    "flake8-eradicate",
    "flake8-isort>=4.0.0",
    "flake8-fixme",
    "flake8-pep3101",
    "flake8-pie",
    "flake8-print",
    "flake8-printf-formatting",
    "flake8-pytest",
    "flake8-pytest-style",
    "flake8-simplify",
    "flake8-todo",
    "flake8-use-fstring",
    "flake8-variables-names",
    "flake8-walrus",
    "flake8-pyproject",

    "mypy",
    "django-stubs",
    "djangorestframework-stubs",
    "types-freezegun",
    "types-Pillow",
]


[tool.flake8]
max-line-length = 160
inline-quotes = "\""
ignore = [
    "DJ05",  # URLs include() should set a namespace
    "E501",  # Line too long
    "E265",  # Block comments should have one space before the pound sign (#) and the comment itself
    "F811",  # Redefinition of unused name from line n
    "PT001",  # Use @pytest.fixture() over @pytest.fixture
    "SIM102",  # Use a single if-statement instead of nested if-statements
    "SIM113",  # Use enumerate instead of manually incrementing a counter
    "E203",  # whitespace before ':', disabled for black purposes https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html#slices
]
exclude = [
    "static",
    "migrations",
    "frontend",
    ".git",
    "__pycache__",
]


[tool.isort]
profile = "google"
line_length = 160
extra_standard_library = ["pytest"]
known_django = ["django", "restframework"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "FIRSTPARTY", "LOCALFOLDER"]
use_parentheses = true
include_trailing_comma = true
multi_line_output = 3


[tool.black]
exclude = '''
/(
  | migrations
)/
'''
line_length = 160


[tool.pytest.ini_options]
 DJANGO_SETTINGS_MODULE = "app.settings"
 python_files = ["test*.py"]
 addopts = ["--reuse-db"]
 markers = [
   "freeze_time: freezing time marker (pytest-freezegun does not register it)",
 ]
 filterwarnings = [  # Pattern: `action:message:category:module:line` (https://docs.python.org/3/library/warnings.html#describing-warning-filters)
   "ignore:.*'rest_framework_jwt.blacklist' defines default_app_config.*You can remove default_app_config.::django",
   "ignore:distutils Version classes are deprecated. Use packaging.version instead.:DeprecationWarning:pytest_freezegun:17",
 ]
 env = [
   "CI=1",
   "CELERY_ALWAYS_EAGER=True",
   "DISABLE_THROTTLING=True",
   "AXES_ENABLED=False",
 ]
